+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 * New cop `LiteralInCondition` tracks uses of literals in the conditions of `if/while/until`.
 * New cop `BeginBlock` tracks uses of `BEGIN` blocks.
 * New cop `EndBlock` tracks uses of `END` blocks.
+* New cop `DotPosition` tracks the dot position in multi-line method calls.
 * Add support for auto-correction of some offences with `-a`/`--auto-correct`.
 
 ### Changes